requirements.txt
Babel~=2.14.0
requests~=2.31.0
aiogram~=3.4.1
mysql-connector-python~=8.3.0
envparse~=0.2.0
python-dotenv~=1.0.1
aiohttp~=3.9.3
aiofiles~=23.2.1
ffmpeg-python~=0.2.0